LayoutTests/imported/w3c/web-platform-tests/svg/animations/scripted/keypoints-attribute-trailing-semi-expected.txt


FAIL A trailing semicolon is allowed in the 'keyPoints' attribute assert_equals: expected -100 but got 0
